Move tests around for legibility, move spectrum processor test to spectrum_reader file
